Fixed minor issues
